+++ LibC/src/sys/time.c
@@ -8,10 +8,10 @@ struct timeval{
 };
 
 int gettimeofday (struct timeval* tv, void* tz){
-	uint32_t sec;
-	uint32_t msec;
+	uint64_t sec;
+	uint64_t msec;
 
-	syscall(SYS_UPTIME,(uint32_t)&sec,(uint32_t)&msec,0,0,0);
+	syscall(SYS_UPTIME,(uint64_t)&sec,(uint64_t)&msec,0,0,0);
 
 	tv->tv_sec = sec;
 	tv->tv_usec = msec*100;
